Google Colab script for translating projects
Note that this script is not perfect, it worked for over 700 projects but failed for ~50. This seems to be due to the special characters present in some languages breaking the XML. The script currently implements storing failed projects to a folder for manual review, but could be improved with better character parsing and encoding of special characters to prevent this failure mode.