Translate projects and upload to Label Studio
This script allows users to translate Label Studio projects into various languages then upload them to Label Studio via the API. It has error handling built in to save any failed project uploads as txt files so they can be manually re-examined and uploaded. The most common failure mode is the length of title (must be under 50 characters), so if you run into errors, suggest starting with finding shorter translations for the the titles of projects.